requirements.txt
tensorflow==2.16.1
numpy==1.25.2
fastapi
uvicorn
pillow
python-multipart